+++ rkt/config/paths.go
@@ -17,10 +17,12 @@ package config
 import (
 	"encoding/json"
 	"fmt"
+	"path/filepath"
 )
 
 type configurablePathsV1 struct {
-	Data string `json:"data"`
+	Data         string `json:"data"`
+	Stage1Images string `json:"stage1-images"`
 }
 
 func init() {
@@ -36,11 +38,21 @@ func (p *configurablePathsV1) parse(config *Config, raw []byte) error {
 	}
 	if dirs.Data != "" {
 		if config.Paths.DataDir != "" {
-			// A clash has occurred. Data dir has been defined more than once in
-			// the same directory
 			return fmt.Errorf("data directory is already specified")
 		}
+		if !filepath.IsAbs(dirs.Data) {
+			return fmt.Errorf("data directory must be an absolute path")
+		}
 		config.Paths.DataDir = dirs.Data
+	}
+	if dirs.Stage1Images != "" {
+		if config.Paths.Stage1ImagesDir != "" {
+			return fmt.Errorf("stage1 images directory is already specified")
+		}
+		if !filepath.IsAbs(dirs.Stage1Images) {
+			return fmt.Errorf("stage1 images directory must be an absolute path")
+		}
+		config.Paths.Stage1ImagesDir = dirs.Stage1Images
 	}
 
 	return nil